Add several new columns to the Credit Application dashboard
Adds descriptive name column to the WFRS credentials model. Then, adds several
new columnes to the application search index and dashboard table:

1. Merchant Name: the name of the credentials used to apply for credit.
2. Application Source: a string provided by the application API to denote where
the applied from. For example: in-checkout, by phone, in a store, etc. Its up to
the API client to provide this string.
3. Requested Value: the requested credit amount given to the application.
4. Credit Limit: if the application was approved, this is the amount of credit
the customer was approved for.
5. Order Total: using email address and the last 4 digits of the account number,
attempt to find the first order placed with the credit line resulting from this
order. If an order is identified, this is the total_incl_tax of that order.
6. Order Merchant Name: Using the same method as column 5 for finding an order,
this is the name of the WFRS credentials used when placing the order.