Refs #13503 Fix merchant ID shown in Pre-Qual Dashboard
Now that the Pre-Qual SDK supports passing in a merchant ID on the client-side,
the MID we were tracking and showing in the dashboard was no longer accurate (as
it was based on the APICredentials which would be used for a SOAP transaction).

This change adds:

1. A new model for entering SDK merchant IDs for the front-end SDK
to use
2. An API for retrieinvg hte SDK merchant ID that should be used
3. New fields on the PreQualificationRequest model to track the merchant ID and
name used by the SDK for a particular request.